+++ pkg/gitqlite/gitqlite_test.go
@@ -2,14 +2,11 @@ package gitqlite
 
 import (
 	"database/sql"
-	"io"
 	"io/ioutil"
 	"os"
 	"testing"
 
 	"github.com/go-git/go-git/v5"
-	"github.com/go-git/go-git/v5/plumbing"
-	"github.com/go-git/go-git/v5/plumbing/object"
 )
 
 var (
@@ -63,356 +60,7 @@ func TestModuleInitialization(t *testing.T) {
 	}
 }
 
-func TestCommitCounts(t *testing.T) {
-	instance, err := New(fixtureRepoDir, &Options{})
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	headRef, err := fixtureRepo.Head()
-	if err != nil {
-		t.Fatal(err)
-	}
-	commitChecker, err := fixtureRepo.Log(&git.LogOptions{
-		From:  headRef.Hash(),
-		Order: git.LogOrderCommitterTime,
-	})
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	commitCount := 0
-	err = commitChecker.ForEach(func(c *object.Commit) error {
-		commitCount++
-		return nil
-	})
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	//checks commits
-	rows, err := instance.DB.Query("SELECT * FROM commits")
-	if err != nil {
-		t.Fatal(err)
-	}
-	defer rows.Close()
-	columns, err := rows.Columns()
-	if err != nil {
-		t.Fatal(err)
-	}
-	expected := 14
-	if len(columns) != expected {
-		t.Fatalf("expected %d columns, got: %d", expected, len(columns))
-	}
-	numRows := getRowsCount(rows)
-
-	expected = commitCount
-	if numRows != expected {
-		t.Fatalf("expected %d rows got: %d", expected, numRows)
-	}
-
-	rows, err = instance.DB.Query("SELECT id, author_name FROM commits")
-	if err != nil {
-		t.Fatal(err)
-	}
-	rowNum, contents, err := getContents(rows)
-	if err != nil {
-		t.Fatalf("err %d at row Number %d", err, rowNum)
-	}
-	for i, c := range contents {
-		commit, err := commitChecker.Next()
-		if err != nil {
-			if err == io.EOF {
-				break
-			} else {
-				t.Fatal(err)
-			}
-		}
-		if commit.ID().String() != c[0] {
-			t.Fatalf("expected %s at row %d got %s", commit.ID().String(), i, c[0])
-		}
-		if commit.Author.Name != c[1] {
-			t.Fatalf("expected %s at row %d got %s", commit.Author.Name, i, c[1])
-		}
-
-	}
-
-}
-
-func TestGoGitCommits(t *testing.T) {
-	instance, err := New(fixtureRepoDir, &Options{SkipGitCLI: true})
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	headRef, err := fixtureRepo.Head()
-	if err != nil {
-		t.Fatal(err)
-	}
-	commitChecker, err := fixtureRepo.Log(&git.LogOptions{
-		From:  headRef.Hash(),
-		Order: git.LogOrderCommitterTime,
-	})
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	commitCount := 0
-	err = commitChecker.ForEach(func(c *object.Commit) error {
-		commitCount++
-		return nil
-	})
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	//checks commits
-	rows, err := instance.DB.Query("SELECT * FROM commits")
-	if err != nil {
-		t.Fatal(err)
-	}
-	defer rows.Close()
-	columns, err := rows.Columns()
-	if err != nil {
-		t.Fatal(err)
-	}
-	expected := 14
-	if len(columns) != expected {
-		t.Fatalf("expected %d columns, got: %d", expected, len(columns))
-	}
-	numRows := getRowsCount(rows)
-
-	expected = commitCount
-	if numRows != expected {
-		t.Fatalf("expected %d rows got: %d", expected, numRows)
-	}
-
-	rows, err = instance.DB.Query("SELECT id, author_name FROM commits")
-	if err != nil {
-		t.Fatal(err)
-	}
-	rowNum, contents, err := getContents(rows)
-	if err != nil {
-		t.Fatalf("err %d at row Number %d", err, rowNum)
-	}
-	for i, c := range contents {
-		commit, err := commitChecker.Next()
-		if err != nil {
-			if err == io.EOF {
-				break
-			} else {
-				t.Fatal(err)
-			}
-		}
-		if commit.ID().String() != c[0] {
-			t.Fatalf("expected %s at row %d got %s", commit.ID().String(), i, c[0])
-		}
-		if commit.Author.Name != c[1] {
-			t.Fatalf("expected %s at row %d got %s", commit.Author.Name, i, c[1])
-		}
-
-	}
-
-}
-
-func TestFileCounts(t *testing.T) {
-	instance, err := New(fixtureRepoDir, &Options{})
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	headRef, err := fixtureRepo.Head()
-	if err != nil {
-		t.Fatal(err)
-	}
-	commitChecker, err := fixtureRepo.Log(&git.LogOptions{
-		From:  headRef.Hash(),
-		Order: git.LogOrderCommitterTime,
-	})
-	if err != nil {
-		t.Fatal(err)
-	}
-	numFiles := 0
-	err = commitChecker.ForEach(func(c *object.Commit) error {
-		numFiles++
-		return nil
-	})
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	fileRows, err := instance.DB.Query("SELECT DISTINCT commit_id FROM files")
-	if err != nil {
-		t.Fatal(err)
-	}
-	defer fileRows.Close()
-	numFileRows := getRowsCount(fileRows)
-	if numFileRows != numFiles {
-		t.Fatalf("expected %d rows got : %d", numFiles, numFileRows)
-	}
-	columnQuery, err := instance.DB.Query("SELECT * FROM files LIMIT 1")
-	if err != nil {
-		t.Fatal(err)
-	}
-	columns, err := columnQuery.Columns()
-	if err != nil {
-		t.Fatal(err)
-	}
-	if len(columns) != 6 {
-		t.Fatalf("expected %d columns got : %d", 6, len(columns))
-	}
-}
-
-func TestRefCounts(t *testing.T) {
-	instance, err := New(fixtureRepoDir, &Options{})
-	if err != nil {
-		t.Fatal(err)
-	}
-	refChecker, err := fixtureRepo.References()
-	if err != nil {
-		t.Fatal(err)
-	}
-	refCount := 0
-	err = refChecker.ForEach(func(r *plumbing.Reference) error {
-		refCount++
-		return nil
-	})
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	//check refs
-	refRows, err := instance.DB.Query("SELECT * FROM refs")
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	columns, err := refRows.Columns()
-	if err != nil {
-		t.Fatal(err)
-	}
-	expected := 3
-	if len(columns) != expected {
-		t.Fatalf("expected %d columns, got: %d", expected, len(columns))
-	}
-	numRows := getRowsCount(refRows)
-	if numRows != refCount {
-		t.Fatalf("expected %d rows got : %d", refCount, numRows)
-	}
-	refRows, err = instance.DB.Query("SELECT * FROM refs")
-	if err != nil {
-		t.Fatal(err)
-	}
-	rowNum, contents, err := getContents(refRows)
-	if err != nil {
-		t.Fatalf("err %d at row Number %d", err, rowNum)
-	}
-	for i, c := range contents {
-		ref, err := refChecker.Next()
-		if err != nil {
-			if err == io.EOF {
-				break
-			} else {
-				t.Fatal(err)
-			}
-		}
-		if ref.Name().String() != c[0] {
-			t.Fatalf("expected %s at row %d got %s", ref.Name().String(), i, c[0])
-		}
-		if ref.Type().String() != c[1] {
-			t.Fatalf("expected %s at row %d got %s", ref.Type().String(), i, c[1])
-		}
-		if ref.Hash().String() != c[2] {
-			t.Fatalf("expected %s at row %d got %s", ref.Hash().String(), i, c[2])
-		}
-
-	}
-}
-
-func TestTags(t *testing.T) {
-	instance, err := New(fixtureRepoDir, &Options{})
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	tagIterator, err := fixtureRepo.Tags()
-	if err != nil {
-		t.Fatal(err)
-	}
-	tagRows, err := instance.DB.Query("SELECT * FROM tags")
-	if err != nil {
-		t.Fatal(err)
-	}
-	rowNum, contents, err := getContents(tagRows)
-	if err != nil {
-		t.Fatalf("err %d at row Number %d", err, rowNum)
-	}
-	for i, c := range contents {
-		tag, err := tagIterator.Next()
-		if err != nil {
-			if err == io.EOF {
-				break
-			} else {
-				t.Fatal(err)
-			}
-		}
-		if tag.Hash().String() != c[0] {
-			t.Fatalf("expected %s at row %d got %s", tag.Hash().String(), i, c[0])
-		}
-		if tag.Name().String() != c[1] {
-			t.Fatalf("expected %s at row %d got %s", tag.Name(), i, c[1])
-		}
-
-	}
-}
-func TestBranches(t *testing.T) {
-	instance, err := New(fixtureRepoDir, &Options{})
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	localBranchIterator, err := fixtureRepo.Branches()
-	if err != nil {
-		t.Fatal(err)
-	}
-	remoteBranchIterator, err := remoteBranches(fixtureRepo.Storer)
-	if err != nil {
-		t.Fatal(err)
-	}
-	branchRows, err := instance.DB.Query("SELECT * FROM branches")
-	if err != nil {
-		t.Fatal(err)
-	}
-	rowNum, contents, err := getContents(branchRows)
-	if err != nil {
-		t.Fatalf("err %d at row Number %d", err, rowNum)
-	}
-	for i, c := range contents {
-		branch, err := localBranchIterator.Next()
-		if err != nil {
-			if err == io.EOF {
-				branch, err = remoteBranchIterator.Next()
-				if err != nil {
-					if err == io.EOF {
-						break
-					} else {
-						t.Fatal(err)
-					}
-				}
-			} else {
-				t.Fatal(err)
-			}
-		}
-		if branch.Name().Short() != c[0] || branch.Hash().String() != c[4] {
-			t.Fatalf("expected %s at row %d got %s", branch.Name().String(), i, c[0])
-		}
-		if branch.Hash().String() != c[4] {
-			t.Fatalf("expected %s at row %d got %s", branch.Hash().String(), i, c[4])
-		}
-
-	}
-}
-func getRowsCount(rows *sql.Rows) int {
+func GetRowsCount(rows *sql.Rows) int {
 	count := 0
 	for rows.Next() {
 		count++
@@ -420,7 +68,7 @@ func getRowsCount(rows *sql.Rows) int {
 
 	return count
 }
-func getContents(rows *sql.Rows) (int, [][]string, error) {
+func GetContents(rows *sql.Rows) (int, [][]string, error) {
 	var (
 		count int = 0
 	)
